refactor: reorganize tool-related code into separate file and clean up unused definitions
